move load plot options
