+++ biofefi/services/plotting.py
@@ -11,34 +11,6 @@ import shap
 from biofefi.options.plotting import PlottingOptions
 
 
-def save_plot_options(path: Path, options: PlottingOptions):
-    """Save plot options to a `json` file at the specified path.
-
-    Args:
-        path (Path): The path to the `json` file.
-        options (PlottingOptions): The options to save.
-    """
-    options_json = dataclasses.asdict(options)
-    with open(path, "w") as json_file:
-        json.dump(options_json, json_file)
-
-
-def load_plot_options(path: Path) -> PlottingOptions:
-    """Load plotting options from the given path.
-    The path will be to a `json` file containing the plot options.
-
-    Args:
-        path (Path): The path the `json` file containing the options.
-
-    Returns:
-        PlottingOptions: The plotting options.
-    """
-    with open(path, "r") as json_file:
-        options_json = json.load(json_file)
-    options = PlottingOptions(**options_json)
-    return options
-
-
 def plot_lime_importance(
     df: pd.DataFrame,
     plot_opts: PlottingOptions,
